Move test common/index.js one level up
