+++ Test/TestCases/napi-dotnet/async_methods.js
@@ -1,8 +1,8 @@
 const assert = require('assert');
-const common = require('../node-addon-api/common');
+const common = require('../common');
 
 /** @type {import('./napi-dotnet')} */
-const binding = require('../node-addon-api/common').binding;
+const binding = common.binding;
 
 common.runTest(async () => {
   const result = await binding.async_method('buddy');
