+++ Test/TestCases/napi-dotnet/hello.js
@@ -1,7 +1,7 @@
 const assert = require('assert');
 
 /** @type {import('./napi-dotnet')} */
-const binding = require('../node-addon-api/common').binding;
+const binding = require('../common').binding;
 
 // Call a method exported by the addon module.
 const result = binding.hello('world');
