+++ Test/TestCases/napi-dotnet/multi_instance.js
@@ -1,12 +1,11 @@
 const assert = require('assert');
 const { Worker, isMainThread, parentPort } = require('worker_threads');
 
+/** @typedef {import('./napi-dotnet')} Binding */
+/** @type Binding */
+const binding = require('../common').binding;
 
-  /** @typedef {import('./napi-dotnet')} Binding */
-  /** @type Binding */
-const  binding = require('../node-addon-api/common').binding;
-
-const { dotnetHost, dotnetModule } = require('../node-addon-api/common');
+const { dotnetHost, dotnetModule } = require('../common');
 
 if (isMainThread) {
   // Increment the static counter to 2.
