+++ Test/TestCases/node-addon-api/basic_types/number.js
@@ -3,7 +3,7 @@
 
 const assert = require('assert');
 
-module.exports = require('../common').runTest(test);
+module.exports = require('../../common').runTest(test);
 
 function test(binding) {
   const MIN_INT32 = -2147483648;
