+++ Test/TestCases/node-addon-api/basic_types/value.js
@@ -2,7 +2,7 @@
 
 const assert = require('assert');
 
-module.exports = require('../common').runTest(test);
+module.exports = require('../../common').runTest(test);
 
 function test(binding) {
   const externalValue = binding.basicTypesValue.createExternal();
