+++ Test/TestCases/node-addon-api/object/finalizer.js
@@ -3,7 +3,7 @@
 const assert = require('assert');
 const testUtil = require('../testUtil');
 
-module.exports = require('../common').runTest(test);
+module.exports = require('../../common').runTest(test);
 
 function createWeakRef (binding, bindingToTest) {
   return binding.object[bindingToTest]({});
