+++ Test/TestCases/node-addon-api/object/get_property.js
@@ -2,7 +2,7 @@
 
 const assert = require('assert');
 
-module.exports = require('../common').runTest(test);
+module.exports = require('../../common').runTest(test);
 
 function test(binding) {
   function testGetProperty(nativeGetProperty) {
